requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==4.2.2
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
cryptography==37.0.1
cycler==0.11.0
dataframe-image==0.1.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
exchange-calendars==3.6.3
fake-useragent==0.1.11
fastjsonschema==2.15.3
fonttools==4.32.0
frozenlist==1.3.0
h11==0.13.0
idna==3.3
importlib-resources==5.6.0
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython_genutils==0.2.0
jedi @ file:///C:/ci/jedi_1644315425835/work
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.0
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mistune==0.8.4
multidict==6.0.2
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
pycparser==2.21
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyluach==2.0.0
pymongo==4.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pyzmq==22.3.0
requests==2.27.1
schedule==1.1.0
selenium==4.1.5
singleton-decorator==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
tabulate==0.8.9
testpath==0.6.0
toolz==0.11.2
tornado==6.1
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webdriver-manager==3.5.4
webencodings==0.5.1
wincertstore==0.2
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0
